requirements.txt
# KitForge - Python Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0

# 3D Processing
trimesh==4.1.3
numpy==1.26.3
pillow==10.2.0
networkx==3.2.1

# Data Models & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# PDF Generation
reportlab==4.0.9
markdown==3.5.2

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
